pyproject.toml
[tool.poetry]
name = "animevisualised"
version = "0.1.0"
description = "A website to make beautiful and insightful graphs on your anime lists."
authors = ["Shravan Asati <dev.shravan@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.1"
flask-sqlalchemy = "^3.1.1"
mysql-connector-python = "^8.3.0"
python-dotenv = "^1.0.1"
flask-login = "^0.6.3"
requests = "^2.31.0"
pandas = "^2.2.0"
numpy = "^1.26.4"
matplotlib = "^3.8.2"
animec = {git = "https://github.com/shravanasati/animec.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
